Revert "seçil"
This reverts commit c07d7d62c5e7a22088e149a0ba5b7c51a16146ce.

+++ src/main/java/com/grup_7/LibraryApp/repository/StaffRepository.java
@@ -1,10 +1,10 @@
 package com.grup_7.LibraryApp.repository;
 
+
 import com.grup_7.LibraryApp.entity.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StaffRepository extends JpaRepository<Staff, Long> {
-
+public interface StaffRepository extends JpaRepository<Staff, Integer> {
 }
